Revert "Authorization added, servers unified"
