+++ .gitignore
@@ -3,8 +3,6 @@ logs
 *.log
 npm-debug.log*
 
-dist
-
 package-lock.json
 
 # Runtime data
