+++ auth/AuthService.js
@@ -1,57 +1,44 @@
-const bcrypt = require('bcrypt');
-const User   = require('../user/model/User');
-const TokenService = require('./TokenService');
+// const bcrypt = require('bcrypt');
+// const User   = require('../user/model/User');
+// const TokenService = require('./TokenService');
 
 
-const isValidUser = async ({ username, password:textPassword }) => {
-  try {
-    // find the user by username
-    const user = await User.findOne(username);
-    console.log(user)
+// const isValidUser = async ({ username, password:textPassword }) => {
+//   try {
+//     // find the user by username
+//     const user = await User.findOne(username);
 
-    // compare the cleartext password and the pwd from the db
-    // using bcrypt
-    const goodPassword = await bcrypt.compare(textPassword, user.password)
-    console.log({goodPassword});
-    return goodPassword;
-  } catch (err) {
-    console.error(err);
-    return false;
-  }
-};
+//     // compare the cleartext password and the pwd from the db
+//     // using bcrypt
+//     //return await bcrypt.compare(textPassword, user.password);
+//   } catch (err) {
+//     console.error(err);
+//     return false;
+//   }
+// };
 
-module.exports = {
-  authenticate(req, res, next) {
-    if (!isValidUser(req.body)) {
-      return next({});
-    }
+// module.exports = {
+//   async authenticate(req, res, next) {
+//     if (await isValidUser(req.body)) {
+//       res.locals.token = await TokenService.makeToken({
+//         username: req.body.username,
+//         roles: ['admin', 'editor'],
+//       });
+//       return next();
+//     }
 
-    TokenService.makeToken({
-        username: req.body.username,
-        roles: ['admin', 'editor'],
-      }).then((token) => {
-         res.locals.token = token;
-         next();
-      }).catch((err) => {
-        debugger;
-        next(err);
-      });
-      return false
-  },
+//     return next({});
+//   },
 
-  allow({roles}) {
-    return [
-      (req, res, next) => {
-        TokenService.verify(req.authToken)
-        .then((payload) => {
-          const isAuthorized = roles.some (n => payload.roles.includes(n));
-          return isAuthorized ? next() : Promise.reject('User not authorized');
-        })
-        .catch(next);
-      },
-      (err, req, res, next) => res.status(403).json({})
-    ];
-  }
-};
+//   isAuth(roles) {
+//     return [
+//       async function isAuth(req, res, next) {
+//         const token = req.headers.authorization;
+//         const isValid = await TokenService.verify(token)
+//         debugger;
+//       },
+//       (err, req, res, next) => res.status(403).json({})
+//     ];
+//   },
 
-
+// };
