+++ client/src/calculateResult.js
