+++ client/src/components/partials/Footer.js
