+++ client/src/components/partials/History.js
@@ -1,0 +1,3 @@
+import createHistory from 'history/createBrowserHistory';
+
+export default createHistory();
