+++ client/src/components/partials/Score.js
