+++ client/src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
-
+import registerServiceWorker from './registerServiceWorker';
 
 import { BrowserRouter as Router } from 'react-router-dom';
 
@@ -9,6 +9,6 @@ ReactDOM.render(
   <Router>
   <App />
   </Router>,
-  document.getElementById('container'));
+  document.getElementById('root'));
 
-
+registerServiceWorker();
