+++ package.json
@@ -1,81 +1,35 @@
 {
-"name": "project3",
-"version": "1.0.0",
-"description": "GA WDI group project-3",
-"main": "server.js",
-
-"dependencies": {
+  "name": "project3",
+  "version": "1.0.0",
+  "description": "GA WDI group project-3",
+  "main": "server.js",
+  "dependencies": {
+    "algorithmia": "^0.3.10",
     "bcrypt": "^1.0.3",
     "body-parser": "^1.18.1",
     "cors": "^2.8.4",
-    "dotenv": "^4.0.0",
     "express": "^4.15.4",
-    "faker": "^4.1.0",
     "jsonwebtoken": "^8.0.1",
     "method-override": "^2.3.9",
     "morgan": "^1.8.2",
     "path": "^0.12.7",
     "pg-promise": "^6.7.1"
   },
-
-"devDependencies": {
-    "algorithmia": "^0.3.10",
-    "axios": "^0.16.2",
-    "babel-core": "^6.26.0",
-    "babel-loader": "^7.1.2",
-    "babel-plugin-transform-object-rest-spread": "^6.26.0",
-    "babel-preset-env": "^1.6.0",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-0": "^6.24.1",
-    "babel-regenerator-runtime": "^6.5.0",
-    "clean-webpack-plugin": "^0.1.16",
-    "css-loader": "^0.28.7",
-    "eslint": "^4.7.2",
-    "eslint-config-airbnb": "^15.1.0",
+  "devDependencies": {
+    "eslint": "^4.7.1",
     "eslint-config-airbnb-base": "^12.0.0",
     "eslint-plugin-import": "^2.7.0",
-    "eslint-plugin-jsx-a11y": "^6.0.2",
-    "eslint-plugin-react": "^7.4.0",
-    "extract-text-webpack-plugin": "^3.0.0",
-    "file-loader": "^0.11.2",
-    "font-awesome": "^4.7.0",
-    "html-webpack-plugin": "^2.30.1",
-    "html-webpack-template": "^5.6.0",
-    "node-sass": "^4.5.3",
-    "nodemon": "^1.12.1",
-    "react": "^15.6.1",
-    "react-dom": "^15.6.1",
-    "react-router": "^4.2.0",
-    "react-router-dom": "^4.2.2",
-    "react-scripts": "^1.0.13",
-    "sass-loader": "^6.0.6",
-    "style-loader": "^0.18.2",
-    "url-loader": "^0.5.9",
-    "webpack": "^3.6.0",
-    "webpack-bundle-analyzer": "^2.9.0"
+    "nodemon": "^1.12.1"
   },
-
-"scripts": {
+  "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "node server.js",
-    "dev": "nodemon server.js",
-    "debug": "nodemon --inspect server.js",
-    "rebuild": "webpack -d --progress",
-    "watch": "webpack -d --progress --watch",
-    "build": "webpack -p"
+    "dev": "nodemon server.js"
   },
-
-"repository": {
+  "repository": {
     "type": "git",
-    "url": "git@github.com:Project3ZNR/Project3.git"
+    "url": "git@git.generalassemb.ly:Zarrina/Project3.git"
   },
-
-"contributors": "Zarrina, Natasha, Richard",
-
-"license": "ISC",
-
-"engines": {"node": ">=8.5.0"}
-
+  "author": "Zarrina, Natasha, Richard",
+  "license": "ISC"
 }
-
